+++ IsekaiMod/Content/Deities/AdministratorD.cs
@@ -12,14 +12,15 @@ using Kingmaker.UnitLogic.FactLogic;
 
 namespace IsekaiMod.Content.Deities
 {
-    class TruckKun
+    class AdministratorD
     {
         public static void Add()
         {
-            var ArtificeDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("9656b1c7214180f4b9a6ab56f83b92fb");
+            var ChaosDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("8c7d778bc39fec642befc1435b00f613");
+            var EvilDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("351235ac5fc2b7e47801f63d117b656c");
+            var KnowledgeDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("443d44b3e0ea84046a9bf304c82a0425");
+            var DeathDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("a099afe1b0b32554199b230699a69525");
             var DestructionDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("6832681c9a91bf946a1d9da28c5be4b4");
-            var ReposeDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("076ba1e3a05fac146acfc956a9f41e95");
-            var TravelDomainAllowed = Resources.GetBlueprint<BlueprintFeature>("c008853fe044bd442ae8bd22260592b7");
 
             var CrusaderSpellbook = Resources.GetBlueprint<BlueprintSpellbook>("673d39f7da699aa408cdda6282e7dcc0");
             var ClericSpellbook = Resources.GetBlueprint<BlueprintSpellbook>("4673d19a0cf2fab4f885cc4d1353da33");
@@ -30,31 +31,33 @@ namespace IsekaiMod.Content.Deities
             var WarpriestClass = Resources.GetBlueprint<BlueprintCharacterClass>("30b5e47d47a0e37438cc5a80c96cfb99");
 
             var FeralChampionArchetype = Resources.GetBlueprint<BlueprintArchetype>("f68ca492c9c15e241ab73735fbd0fb9f");
+            var PriestOfBalance = Resources.GetBlueprint<BlueprintArchetype>("a4560e3fb5d247d68fb1a2738fcc0855");
             var AngelfireApostle = Resources.GetBlueprint<BlueprintArchetype>("857bc9fadf70f294795a9cba974a48b8");
 
-            var ChannelPositiveAllowed = Resources.GetBlueprint<BlueprintFeature>("8c769102f3996684fb6e09a2c4e7e5b9");
             var ChannelNegativeAllowed = Resources.GetBlueprint<BlueprintFeature>("dab5255d809f77c4395afc2b713e9cd6");
 
-            var ShieldsProficiency = Resources.GetBlueprint<BlueprintFeature>("cb8686e7357a68c42bdd9d4e65334633");
-            var ShieldBashFeature = Resources.GetBlueprint<BlueprintFeature>("121811173a614534e8720d7550aae253");
-            var HeavyShieldPlus1 = Resources.GetBlueprint<BlueprintItem>("5c7b898a1bfb6cb4f8c14a0ebc143abe");
+            var ScytheProficiency = Resources.GetBlueprint<BlueprintFeature>("96c174b0ebca7b246b82d4bc4aac4574");
+            var ScythePlus1 = Resources.GetBlueprint<BlueprintItem>("8933943621eca2d45b99d851bd9100d9");
 
-            var Icon_Truck = AssetLoader.LoadInternal("Deities", "ICON_TRUCK.png");
-            var TruckKunFeature = Helpers.CreateBlueprint<BlueprintFeature>("TruckKunFeature", (bp => {
-                bp.SetName("Truck-kun");
+            var Icon_AdministratorD = AssetLoader.LoadInternal("Deities", "ICON_ADMIN_D.png");
+            var AdministratorDFeature = Helpers.CreateBlueprint<BlueprintFeature>("AdministratorDFeature", (bp => {
+                bp.SetName("Administrator D");
                 bp.SetDescription(
-                    "Truck-kun is a menace of the streets, a killer that prowls the road before it smashes their victim so hard they get transported into another world. "
-                    + "Many fear death brought by Truck-kun, but there are some who believe that their reincarnation will bring them into a harem paradise. "
-                    + "Beware of the streets, and look at both sides of the road, or else you might be its latest victim."
-                    + "\nDomains: Artifice, Destruction, Repose, Travel"
-                    + "\nFavoured Weapon: Heavy Shield Bash");
-                bp.m_Icon = Icon_Truck;
+                    "Administrator D is a playful, self-proclaimed \"Ultimate evil god\" that does things simply for entertainment. "
+                    + "She is sadistic and enjoys watching the protagonist struggle for survival in the new world for her own amusement."
+                    + "\nDomains: Chaos, Evil, Knowledge, Death, Destruction"
+                    + "\nFavoured Weapon: Scythe");
+                bp.m_Icon = Icon_AdministratorD;
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
                 bp.HideInCharacterSheetAndLevelUp = false;
                 bp.Groups = new FeatureGroup[] { FeatureGroup.Deities };
 
                 // Exclude from Archetypes
+                bp.AddComponent<PrerequisiteNoArchetype>(c => {
+                    c.m_CharacterClass = ClericClass.ToReference<BlueprintCharacterClassReference>();
+                    c.m_Archetype = PriestOfBalance.ToReference<BlueprintArchetypeReference>();
+                });
                 bp.AddComponent<PrerequisiteNoArchetype>(c => {
                     c.m_CharacterClass = ClericClass.ToReference<BlueprintCharacterClassReference>();
                     c.m_Archetype = AngelfireApostle.ToReference<BlueprintArchetypeReference>();
@@ -66,17 +69,17 @@ namespace IsekaiMod.Content.Deities
 
                 // Alignment
                 bp.AddComponent<PrerequisiteAlignment>(c => {
-                    c.Alignment = AlignmentMaskType.LawfulNeutral | AlignmentMaskType.ChaoticNeutral | AlignmentMaskType.TrueNeutral | AlignmentMaskType.NeutralGood | AlignmentMaskType.NeutralEvil;
+                    c.Alignment = AlignmentMaskType.ChaoticEvil | AlignmentMaskType.ChaoticNeutral | AlignmentMaskType.NeutralEvil;
                 });
 
                 // Domains & Energy Channel
                 bp.AddComponent<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] {
-                        ArtificeDomainAllowed.ToReference<BlueprintUnitFactReference>(),
+                        ChaosDomainAllowed.ToReference<BlueprintUnitFactReference>(),
+                        EvilDomainAllowed.ToReference<BlueprintUnitFactReference>(),
+                        KnowledgeDomainAllowed.ToReference<BlueprintUnitFactReference>(),
+                        DeathDomainAllowed.ToReference<BlueprintUnitFactReference>(),
                         DestructionDomainAllowed.ToReference<BlueprintUnitFactReference>(),
-                        ReposeDomainAllowed.ToReference<BlueprintUnitFactReference>(),
-                        TravelDomainAllowed.ToReference<BlueprintUnitFactReference>(),
-                        ChannelPositiveAllowed.ToReference<BlueprintUnitFactReference>(),
                         ChannelNegativeAllowed.ToReference<BlueprintUnitFactReference>()
                     };
                 });
@@ -93,17 +96,7 @@ namespace IsekaiMod.Content.Deities
                 // Cleric, Inquistor, Warpriest starting proficiency and weapon
                 bp.AddComponent<AddFeatureOnClassLevel>(c => {
                     c.m_Class = ClericClass.ToReference<BlueprintCharacterClassReference>();
-                    c.m_Feature = ShieldsProficiency.ToReference<BlueprintFeatureReference>();
-                    c.Level = 1;
-                    c.m_Archetypes = null;
-                    c.m_AdditionalClasses = new BlueprintCharacterClassReference[2] {
-                        InquistorClass.ToReference<BlueprintCharacterClassReference>(),
-                        WarpriestClass.ToReference<BlueprintCharacterClassReference>()
-                    };
-                });
-                bp.AddComponent<AddFeatureOnClassLevel>(c => {
-                    c.m_Class = ClericClass.ToReference<BlueprintCharacterClassReference>();
-                    c.m_Feature = ShieldBashFeature.ToReference<BlueprintFeatureReference>();
+                    c.m_Feature = ScytheProficiency.ToReference<BlueprintFeatureReference>();
                     c.Level = 1;
                     c.m_Archetypes = null;
                     c.m_AdditionalClasses = new BlueprintCharacterClassReference[2] {
@@ -112,7 +105,7 @@ namespace IsekaiMod.Content.Deities
                     };
                 });
                 bp.AddComponent<AddStartingEquipment>(c => {
-                    c.m_BasicItems = new BlueprintItemReference[1] { HeavyShieldPlus1.ToReference<BlueprintItemReference>() };
+                    c.m_BasicItems = new BlueprintItemReference[1] { ScythePlus1.ToReference<BlueprintItemReference>() };
                     c.m_RestrictedByClass = new BlueprintCharacterClassReference[3] {
                         ClericClass.ToReference<BlueprintCharacterClassReference>(),
                         InquistorClass.ToReference<BlueprintCharacterClassReference>(),
@@ -123,7 +116,7 @@ namespace IsekaiMod.Content.Deities
 
             // Add to Deity Selection
             var IsekaiDeitySelection = Resources.GetModBlueprint<BlueprintFeatureSelection>("IsekaiDeitySelection");
-            IsekaiDeitySelection.m_AllFeatures = IsekaiDeitySelection.m_AllFeatures.AddToArray(TruckKunFeature.ToReference<BlueprintFeatureReference>());
+            IsekaiDeitySelection.m_AllFeatures = IsekaiDeitySelection.m_AllFeatures.AddToArray(AdministratorDFeature.ToReference<BlueprintFeatureReference>());
         }
     }
 }